+++ src/app/[user]/[post]/edit/edit-page.tsx
@@ -2,7 +2,7 @@
 
 import { PostDetailedResponse, UpdatePostRequest } from "@/shared/api/openapi";
 import { ErrorMessage } from "@/shared/ui/error-message";
-import { Loading } from "@/shared/ui/loading";
+import { Loading } from "@/shared/ui/feedback/loading";
 import ConfirmDeleteModal from "@/shared/ui/overlays/confirmation-modal";
 import { PostForm } from "@/features/post/ui/post-form";
 import { IsAuthor, ProtectedRoute } from "@/entities/session";
