+++ src/app/[user]/loading.tsx
@@ -1,4 +1,4 @@
-import { Loading } from "@/shared/ui/loading";
+import { Loading } from "@/shared/ui/feedback/loading";
 
 export default function UserLoading() {
   return <Loading />;
