+++ src/shared/ui/feedback/index.ts
@@ -1,0 +1,5 @@
+export { StatusBadge } from "./status-badge";
+export type { Status } from "./status-badge";
+export { InfoBox } from "./info-box";
+export { Loading, Spinner } from "./loading";
+export { ProgressBar } from "./progress-bar";
