+++ src/shared/ui/feedback/info-box.tsx
