+++ src/shared/ui/image-cropper-dynamic.tsx
@@ -1,5 +1,5 @@
 import dynamic from "next/dynamic";
-import { Loading } from "@/shared/ui/loading";
+import { Loading } from "@/shared/ui/feedback/loading";
 
 export const ImageCropper = dynamic(() => import("./image-cropper"), {
   ssr: false,
